# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeDo Backend API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the WeDo backend."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["code", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        },
        "required": ["error"]
      },
...
      "UserSnippet": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "avatarFileId": { "type": "string", "nullable": true },
          "rating": { "type": "number" }
        },
        "required": ["id", "name", "rating"]
      },
...
      "ActivityLocation": {
        "type": "object",
        "properties": {
          "latitude": { "type": "number" },
          "longitude": { "type": "number" },
          "address": { "type": "string" },
          "name": { "type": "string", "nullable": true },
          "settlement": { "type": "string", "nullable": true },
          "region": { "type": "string", "nullable": true },
          "country": { "type": "string", "nullable": true }
        },
        "required": ["latitude", "longitude", "address"]
      },
      "ActivityPreferences": {
        "type": "object",
        "properties": {
          "gender": { "type": "string", "enum": ["male", "female"], "nullable": true },
          "ageFrom": { "type": "integer", "nullable": true },
          "ageTo": { "type": "integer", "nullable": true },
          "level": { "type": "string", "enum": ["beginner", "intermediate", "advanced"], "nullable": true },
          "maxParticipants": { "type": "integer", "nullable": true }
        }
      },
      "ActivityListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time" },
          "format": { "type": "string", "enum": ["online", "offline"] },
          "status": { "type": "string", "enum": ["active", "cancelled"] },
          "location": { "$ref": "#/components/schemas/ActivityLocation" },
          "categoryId": { "type": "string" },
          "subcategoryId": { "type": "string", "nullable": true },
          "coverPhotoFileId": { "type": "string", "nullable": true },
          "photoFileIds": { "type": "array", "items": { "type": "string" } },
          "organizer": { "$ref": "#/components/schemas/UserSnippet" },
          "participantsCount": { "type": "integer" },
          "maxParticipants": { "type": "integer", "nullable": true },
          "requiresApproval": { "type": "boolean" },
          "preferences": { "allOf": [{ "$ref": "#/components/schemas/ActivityPreferences" }], "nullable": true },
          "pendingRequestsCount": { "type": "integer", "nullable": true },
          "price": { "type": "number" }
        },
        "required": ["id", "title", "startAt", "endAt", "format", "status", "location", "categoryId", "photoFileIds", "organizer", "participantsCount", "requiresApproval", "price"]
      },
      "ActivityDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/ActivityListItem" },
          {
            "type": "object",
            "properties": {
              "description": { "type": "string" },
              "timeZone": { "type": "string" },
              "participantsPreview": { "type": "array", "items": { "$ref": "#/components/schemas/UserSnippet" } },
              "isSaved": { "type": "boolean" },
              "participationStatus": { "type": "string", "nullable": true },
              "isFull": { "type": "boolean" },
              "spotsLeft": { "type": "integer", "nullable": true },
              "policyFlags": {
                "type": "object",
                "properties": {
                  "canJoin": { "type": "boolean" },
                  "canLeave": { "type": "boolean" },
                  "canCancelRequest": { "type": "boolean" },
                  "canManageRequests": { "type": "boolean" },
                  "canRate": { "type": "boolean" },
                  "canEdit": { "type": "boolean" },
                  "canCancelActivity": { "type": "boolean" }
                },
                "required": ["canJoin", "canLeave", "canCancelRequest", "canManageRequests", "canRate", "canEdit", "canCancelActivity"]
              }
            },
            "required": ["description", "timeZone", "participantsPreview", "isSaved", "isFull", "policyFlags"]
          }
        ]
      },
      "ActivityListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ActivityListItem" } },
          "nextCursor": { "type": "string", "nullable": true },
          "hasMore": { "type": "boolean" }
        },
        "required": ["items", "hasMore"]
      },
      "CreateActivityRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "categoryId": { "type": "string" },
          "subcategoryId": { "type": "string", "nullable": true },
          "format": { "type": "string", "enum": ["online", "offline"] },
          "location": { "$ref": "#/components/schemas/ActivityLocation" },
          "startAt": { "type": "string", "format": "date-time" },
          "endAt": { "type": "string", "format": "date-time" },
          "timeZone": { "type": "string" },
          "preferences": { "$ref": "#/components/schemas/ActivityPreferences" },
          "requiresApproval": { "type": "boolean" },
          "photoFileIds": { "type": "array", "items": { "type": "string" } },
          "price": { "type": "string" }
        },
        "required": ["title", "description", "categoryId", "format", "location", "startAt", "endAt", "price"]
      },
...
      "File": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "originalName": { "type": "string" },
          "mimeType": { "type": "string" },
          "size": { "type": "integer" }
        },
        "required": ["id", "originalName", "mimeType", "size"]
      },
      "FilesResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/File" } }
        },
        "required": ["items"]
      },
...
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/activities": {
      "get": {
        "tags": ["activities"],
        "summary": "List activities",
        "parameters": [
          { "name": "categoryId", "in": "query", "schema": { "type": "string" } },
          { "name": "subcategoryId", "in": "query", "schema": { "type": "string" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["online", "offline"] } },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "dateFrom", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "dateTo", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "level", "in": "query", "schema": { "type": "string" } },
          { "name": "gender", "in": "query", "schema": { "type": "string" } },
          { "name": "ageFrom", "in": "query", "schema": { "type": "integer" } },
          { "name": "ageTo", "in": "query", "schema": { "type": "integer" } },
          { "name": "requiresApproval", "in": "query", "schema": { "type": "boolean" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Activities page",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["activities"],
        "summary": "Create activity",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateActivityRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created activity",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ActivityDetail" }
              }
            }
          }
        }
      }
    },
...
    "/files": {
      "get": {
        "tags": ["files"],
        "summary": "Get files by ids",
        "parameters": [
          { "name": "ids", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Comma-separated file ids" }
        ],
        "responses": {
          "200": {
            "description": "Files list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FilesResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["files"],
        "summary": "Upload image file",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded file metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/File" }
              }
            }
          }
        }
      }
    },
    "/files/{file_id}": {
      "get": {
        "tags": ["files"],
        "summary": "Get file binary",
        "parameters": [
          { "name": "file_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Binary file content",
            "content": {
              "image/jpeg": { "schema": { "type": "string", "format": "binary" } },
              "image/png": { "schema": { "type": "string", "format": "binary" } },
              "image/webp": { "schema": { "type": "string", "format": "binary" } }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  }
}